% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_avg_1K.R
\name{eval_avg_1K}
\alias{eval_avg_1K}
\title{Evaluate avg treatment effects over event times 1..K}
\usage{
eval_avg_1K(fit, sim, K)
}
\arguments{
\item{fit}{Fit object returned by multiout_synth()}

\item{sim}{Simulation object returned by simulate_multi_outcome_panel()}

\item{K}{Integer. Max event time used.}
}
\value{
List with tau_hat_avg_1K, tau_true_avg_1K, cor_avg_1K, rmse_avg_1K
}
\description{
Evaluate avg treatment effects over event times 1..K
}
