% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_parallel.R
\name{run_parallel}
\alias{run_parallel}
\title{Internal helper: run a function over X in a chosen parallel backend}
\usage{
run_parallel(
  X,
  FUN,
  parallel_backend = c("none", "fork", "psock"),
  n_cores = 2L,
  export = NULL,
  envir = parent.frame()
)
}
\arguments{
\item{X}{Vector/list of inputs.}

\item{FUN}{Function of one argument.}

\item{parallel_backend}{One of "none","fork","psock".}

\item{n_cores}{Integer number of workers/cores.}

\item{export}{Character vector of object names to export (psock only).}

\item{envir}{Environment to export from (psock only).}
}
\value{
List of length length(X).
}
\description{
Internal helper: run a function over X in a chosen parallel backend
}
\keyword{internal}
